+++ core/src/main/java/eu/unicore/ucc/authn/oidc/UCCOIDCServerAuthN.java
@@ -6,6 +6,7 @@ import org.apache.hc.core5.http.HttpMessage;
 
 import eu.unicore.services.restclient.oidc.OIDCProperties;
 import eu.unicore.services.restclient.oidc.OIDCServerAuthN;
+import eu.unicore.ucc.authn.CallbackUtils;
 
 /**
  * Gets a Bearer token from an OIDC server.
@@ -31,10 +32,26 @@ public class UCCOIDCServerAuthN extends TokenBasedAuthN {
 	@Override
 	public void setProperties(Properties properties)
 	{
+		OIDCProperties oidcProperties = new OIDCProperties(properties);
 		super.setProperties(properties);
 		this.auth = new eu.unicore.services.restclient.oidc.OIDCServerAuthN(
-				new OIDCProperties(properties),
-				super.getAnonymousClientConfiguration());	
+				oidcProperties,
+				super.getAnonymousClientConfiguration()){
+				protected void retrieveToken()throws Exception {
+					String password = oidcProperties.getValue(OIDCProperties.PASSWORD);
+					if(password==null){
+						password = new String(CallbackUtils.getPasswordFromUserCmd("OIDC server", null));
+						oidcProperties.setProperty(oidcProperties.PASSWORD, password);
+						properties.setProperty("oidc.password", password);
+					}
+					String otp = oidcProperties.getValue(OIDCProperties.OTP);
+					if(otp!=null && otp.equalsIgnoreCase("QUERY")){
+						oidcProperties.setProperty(oidcProperties.OTP, otp);
+						otp = new String(CallbackUtils.getPasswordFromUserCmd("OIDC server", "2FA one time "));
+					}
+					super.retrieveToken();
+				}
+			};
 	}
 
 	@Override